#1 Took the code out of the techrentalapp file and swapped to the renttab and added obserablelist for auto updates for data changes
